requirements.txt
APScheduler==3.0.3
docutils==0.11
egenix-mx-base==3.2.8
futures==2.2.0
Jinja2==2.7.2
MarkupSafe==0.23
numpydoc==0.4
pexpect==3.2
pika==0.9.8
Pygments==1.6
pymongo==2.7
python-dateutil==2.2
pytz==2015.2
PyYAML==3.11
requests==2.3.0
retrying==1.3.3
six==1.7.0
Sphinx==1.2.2
tzlocal==1.1.3
Unidecode==0.4.16
xmltodict==0.9.0
